Move str accessor tests to test_str
